requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
transformers
torch
pillow
accelerate
sentencepiece
protobuf
diffusers
safetensors
hf_xet